+++ CHANGELOG.md
@@ -23,7 +23,7 @@ _2021/07/29_
 * New helper function `DomUtil.isArray` to test if a JavaScript object is an array
 * New `EntityAccessor` object which allows to get attributes and child elements from xml or json objects, regardless of their representation
 * New `escapeXtk` function to escape litteral values in Xtk expression. Can be used as a function or as a tagged template litteral
-* New function `XtkCaster.variantStorageAttribute` which returns the name of a schema attribute used to store variant value types. The name of the attribute depends on the type: stringValue, longValue, etc.
+* New function `XtkCaster._variantStorageAttribute` which returns the name of a schema attribute used to store variant value types. The name of the attribute depends on the type: stringValue, longValue, etc.
 
 _Breaking changes in 1.0.0_
 * The default representation is now `SimpleJson` instead of `BadgerFish`